feat: part4 terminada

+++ Part4/bloglist/controllers/blogs.js
@@ -1,41 +1,68 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
-// eslint-disable-next-line no-unused-vars
-const logger = require('../utils/logger')
+const User = require('../models/user')
+
+
 
 
 
 
 blogsRouter.get('/', async(request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog.find({}).populate('user',{ username:1,name:1 })
   response.json(blogs)
 })
-/* blogsRouter.get('/:_id',async(request,response) => {
-  const blogs = await Blog.findById(request.params.id)
-  response.json(blogs)
-}) */
+
 
 blogsRouter.post('/', async(request, response) => {
-  const blog = new Blog(request.body)
+
+  if (!request.token || !request.decodedToken) {
+    return response.status(401).json({ error: 'missing or invalid token' })
+  }
+  const user = await User.findById(request.decodedToken.id)
+  const body = request.body
+
+  const blog = new Blog({
+    title: body.title,
+    author: body.author,
+    url: body.url,
+    likes: body.likes,
+    user: user._id,
+  })
+
+
+
+
   if(!blog.likes){
     blog.likes = 0
   }
   if(!blog.title || !blog.url){
     return response.status(400).end()
   }
+
+
   const blogs = await blog.save()
+  user.blogs = user.blogs.concat(blogs._id)
+  await user.save()
   response.status(201).json(blogs)
 })
 blogsRouter.delete('/:id',async(request,response) => {
-  const blog = await Blog.findById(request.params.id)
-  if(!blog){
-    console.log('no existe ese blog')
+  if (!request.token || !request.decodedToken) {
+    return response.status(401).json({ error: 'missing or invalid token' })
   }
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+  const user = await User.findById(request.decodedToken.id)
+  const blog = await Blog.findById(request.params.id)
+
+
+  if (user.id.toString() === blog.user.toString()) {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  } else {
+    return response.status(401).end()
+  }
 })
 blogsRouter.put('/:id',async(request,response) => {
   const body = request.body
+
 
   const blog = {
     title:body.title,
